+++ lqmt/test/test_data/filters/filter_configs.py
@@ -1,0 +1,11 @@
+TESTCONFIG1 = """
+[[Source.Filters]]
+    site_includes = [ 'SiTE1' ]
+    site_excludes = [ 'SitE2' ]
+    payload_types = [ 'alert' ]
+    payload_formats = [ 'STix' ]
+    sensitivities = [ 'nosensITIvity' ]
+    restrictions = [ 'whIte' ]
+    reconnaissance = [ 'notouch' ]
+    max_file_age = "2 mon"
+"""