+++ lqmt/test/test_data/filters/filter_inputs.py
@@ -1,0 +1,9 @@
+INPUT1 = """{ 
+    "SendingSite": "SITE1", 
+    "SentTimestamp": "1502815271", 
+    "PayloadType": "Alert", 
+    "PayloadFormat": "STIX", 
+    "DataSensitivity": "noSensitivity", 
+    "SharingRestrictions": "WHITE", 
+    "ReconPolicy": "noTouch"
+}"""